+++ src/views/dex/detail/dashboard/DexViewDetailContract.vue
