+++ src/views/dex/detail/dashboard/DexViewDetailLp.vue
