+++ src/views/dex/detail/dashboard/DexViewDetailToken.vue
